+++ jassrealtime/search/multicorpus/documents_by_annotation.py
@@ -56,7 +56,7 @@ class DocumentsByAnnotation(DocumentsBy):
                          must_not=grouped_queries["must_not"],
                          should=grouped_queries["should"])
 
-        # Note: using aggregation means we could lose some hits:
+        # Note: using aggregation means we could lose some hits
         # "when there are lots of unique terms, Elasticsearch only returns the top terms"
         # https://www.elastic.co/guide/en/elasticsearch/reference/6.1/search-aggregations-bucket-terms-aggregation.html#CO105-2
 
@@ -64,23 +64,41 @@ class DocumentsByAnnotation(DocumentsBy):
         search.aggs.bucket('doc_ids', aggregate_by_document_id)
         response = search.execute()
         aggregation = response.aggregations['doc_ids']
+        count = len(aggregation.buckets)  # we want the count before slicing
 
         # Note: Elasticsearch DSL slicing API does not work with aggregations, so we slice on aggregation results
         buckets = aggregation.buckets[from_index:from_index+size]
-        count = len(buckets)
-
-        # We will use the number of matching annotations per document as a naive score
-        annotation_buckets = [{'doc_id': bucket['key'], 'score': bucket['doc_count']} for bucket in buckets]
 
         # Get document information from document id
-        # Set of document ids (id uniqueness is already guaranteed by the aggregation)
+        # (Document id uniqueness is guaranteed by the aggregation.)
         document_ids = [bucket['key'] for bucket in buckets]
         documents = self.documents_by_ids(grouped_targets, document_ids)
 
-        # todo Join with score
-        # Can score order be retained by the "join" or should we sort after?
+        documents_with_score = self.join_documents_with_score(buckets, documents)
 
-        return count, documents
+        return count, documents_with_score
+
+    @staticmethod
+    def join_documents_with_score(buckets, documents):
+        """
+        Join documents with score while retaining scoring order.
+
+        We will use the number of matching annotations per document as a naive score.
+
+        The buckets are naturally ordered by descending number of annotations.
+
+        :param buckets:
+        :param documents:
+        :return:
+        """
+        document_map = {document['id']: document for document in documents}
+        documents_with_score = []
+        for bucket in buckets:
+            document = document_map[bucket['key']]
+            score = bucket['doc_count']
+            document['score'] = score
+            documents_with_score.append(document)
+        return documents_with_score
 
     def targets_indices(self, grouped_targets: dict, schema_types: list) -> list:
         return [self.group_indices(corpus_id, buckets_ids, schema_types) for corpus_id, buckets_ids in grouped_targets.items()]